// src/frontend/WhatToCook.WebApp/src/app/api-contracts/model/recipeRequest.ts
/**
 * WhatToCook.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface RecipeRequest {
  id: number;
  image: string;
  ingredients: Array<string>;
  name: string;
  preparationDescription: string;
  timeToPrepare: string;
  tags: Array<string>;
  shortDescription: string;
}